+++ src/bots/slack/controllers/slack-event.controller.spec.ts
@@ -1,8 +1,8 @@
-import { request } from '../../spec/spec-helpers';
+import { request } from '../../../spec/spec-helpers';
 
 describe('slack event controller API', () => {
   describe('Setup Confirmation', () => {
-    it("should respond to Slack's confirmation request with the correct challeng string", async () => {
+    it("should respond to Slack's confirmation request with the correct challenge string", async () => {
       const sampleBody = { challenge: 'this is a string' };
 
       await request