+++ CHANGELOG.md
@@ -13,3 +13,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Docker for local developmment
 - Connect Slackbot
 - Mongo DB and Mongoose
+- Adds Redis
+- Adds Bull for async + scheduled Jobs